Moved ProxyAuthenticator into its own class
